+++ com.ca.lsp.cobol/lsp-service-cobol/src/test/java/com/ca/lsp/cobol/service/CpyRecognizedWithoutConfigTest.java
@@ -1,6 +1,5 @@
 /*
  * Copyright (c) 2020 Broadcom.
- *
  * The term "Broadcom" refers to Broadcom Inc. and/or its subsidiaries.
  *
  * This program and the accompanying materials are made
@@ -10,17 +9,13 @@
  * SPDX-License-Identifier: EPL-2.0
  *
  * Contributors:
- * Broadcom, Inc. - initial API and implementation
- *
- *
+ *    Broadcom, Inc. - initial API and implementation
  */
 
 package com.ca.lsp.cobol.service;
 
 import com.broadcom.lsp.domain.cobol.databus.api.DataBusBroker;
 import com.broadcom.lsp.domain.cobol.event.model.RequiredCopybookEvent;
-import com.ca.lsp.cobol.model.ConfigurationSettingsStorable;
-import com.ca.lsp.cobol.service.delegates.communications.Communications;
 import com.ca.lsp.cobol.service.delegates.dependency.CopybookDependencyServiceImpl;
 import com.google.inject.Provider;
 import lombok.extern.slf4j.Slf4j;
@@ -38,7 +33,6 @@ import java.util.Collections;
 import java.util.Comparator;
 import java.util.List;
 
-import static com.ca.lsp.cobol.service.delegates.communications.CopybookMessageInfo.*;
 import static org.junit.Assert.fail;
 import static org.mockito.Mockito.*;
 
@@ -49,17 +43,17 @@ import static org.mockito.Mockito.*;
  * dependency file without a dependency folder.
  */
 @Slf4j
-public class CopybookSearchWithoutConfigTest {
+public class CpyRecognizedWithoutConfigTest {
   private static final String CPY_NAME = "ACPYTEST";
   private DataBusBroker dataBusBroker = mock(DataBusBroker.class);
   private Provider settingsProvider = mock(Provider.class);
   private CopybookDependencyServiceImpl dependencyService =
       mock(CopybookDependencyServiceImpl.class);
-  private Communications communications = mock(Communications.class);
   private CopybookServiceImpl copybookService;
   private static final String DOCUMENT_URI = "file:///C:/Users/test/Test.cbl";
   private Path workspaceFolderPath = null;
   private RequiredCopybookEvent requiredCopybookEvent;
+
   /** Initialize the copybook service and dependency service but not the relevant folders */
   @Before
   public void init() {
@@ -67,7 +61,7 @@ public class CopybookSearchWithoutConfigTest {
         createFolderStructure(Paths.get(System.getProperty("java.io.tmpdir"), "WORKSPACE"));
 
     copybookService =
-        new CopybookServiceImpl(dataBusBroker, settingsProvider, dependencyService, communications);
+        new CopybookServiceImpl(dataBusBroker, settingsProvider, dependencyService);
     copybookService.setWorkspaceFolders(generateWorkspaceFolder());
 
     requiredCopybookEvent =
@@ -105,7 +99,7 @@ public class CopybookSearchWithoutConfigTest {
    */
   @Test
   public void findCopybookWithoutSettingsDefined() {
-    // simulate that user have not provided settings for profile and datasetlist but the
+    // simulate that user have not provided settings for profile and dataset list but the
     // preprocessor found a copybook.
 
     // when asking for a setting config we'll give an empty object as defined in the API
@@ -114,44 +108,6 @@ public class CopybookSearchWithoutConfigTest {
     try {
       copybookService.observerCallback(requiredCopybookEvent);
       verifyThatDepFileIsCreated();
-      verify(communications, times(1)).notifyCopybookMessageInfo(NO_SETTINGS);
-    } catch (Exception e) {
-      fail(e.getMessage());
-    }
-  }
-
-  /**
-   * This test verify that when the settings configuration contains just the profile name the server
-   * will not crash and a syntax error of copybook not found is shown to the user.
-   */
-  @Test
-  public void findCopybookWithOnlyProfileDefined() {
-    // simulate that user have not provided settings for profile and datasetlist but the
-    // preprocessor found a copybook.
-    when(settingsProvider.get())
-        .thenReturn(new ConfigurationSettingsStorable(Collections.emptyList()));
-
-    try {
-      copybookService.observerCallback(requiredCopybookEvent);
-      verifyThatDepFileIsCreated();
-      verify(communications, times(1)).notifyCopybookMessageInfo(NO_DATASET_IN_SETTINGS);
-    } catch (Exception e) {
-      fail(e.getMessage());
-    }
-  }
-
-  /**
-   * This test will verify that without the definition of a copybook folder the server will not
-   * crash and a syntax error copybook not found is shown to the user
-   */
-  @Test
-  public void findCopybookWithoutCopybookFolderDefined() {
-    when(settingsProvider.get())
-        .thenReturn(new ConfigurationSettingsStorable(Collections.singletonList("HLQ.DS1")));
-    try {
-      copybookService.observerCallback(requiredCopybookEvent);
-      verifyThatDepFileIsCreated();
-      verify(communications, times(1)).notifyLogMessageInfo(COPYBOOK_FOLDER_MISS);
     } catch (Exception e) {
       fail(e.getMessage());
     }